backend/requirements.txt
setuptools>=70.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydub==0.25.1
audioop-lts==0.2.1
requests==2.32.3
openai>=1.35.0
Pillow>=10.0.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyMySQL==1.1.1
openai-whisper==20250625
anthropic>=0.30.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
python-docx>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
tiktoken>=0.7.0
langdetect>=1.0.9
python-dotenv>=1.0.0